+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 3.2.1-alpha.11 (2025-02-02)
 
 * (theimo1221) Dachs to respect block automatic